feature-image-improvements: wip
