+++ app/components/BlogItem.tsx
@@ -1,11 +1,13 @@
 import dayjs from 'dayjs';
-import { Box, Text, Image } from '@chakra-ui/react';
+import { Box, Text } from '@chakra-ui/react';
 import { Link } from '@remix-run/react';
-import { PostOrPage } from '@tryghost/content-api';
 import TopicsList from './TopicsList';
+import { PostImage } from './PostImage';
+import { JsonCompatibleObject } from './types';
+import { GetPostOutput } from '~/content-api/types';
 
 interface BlogListItemProps {
-  post?: PostOrPage;
+  post?: JsonCompatibleObject<GetPostOutput>;
   type: string;
 }
 
@@ -13,7 +15,6 @@ export default function BlogItem({ post, type }: BlogListItemProps) {
   if (!post) return null;
 
   const hasFeatureImage = !!post.feature_image;
-  const featureImageStartsWithSlash = (post.feature_image ?? '').startsWith('/');
 
   return (
     <Box
@@ -31,27 +32,7 @@ export default function BlogItem({ post, type }: BlogListItemProps) {
           width={{ base: '100%', md: type === 'primary' ? '60%' : '100%' }}
         >
           <Link to={`/${post.slug}`}>
-            <Image
-              src={post.feature_image as string}
-              srcSet={
-                featureImageStartsWithSlash
-                  ? `
-                /api/image?name=${post.feature_image}&w=480 480w,
-                /api/image?name=${post.feature_image}&w=800 800w,
-                /api/image?name=${post.feature_image}&w=1200 1200w
-              `
-                  : undefined
-              }
-              sizes={featureImageStartsWithSlash ? '(max-width: 800px) 100vw, 1200px' : undefined}
-              alt={post.feature_image_alt || 'image'}
-              position="absolute"
-              top="50%"
-              left="50%"
-              transform="translate(-50%, -50%)"
-              width="100%"
-              height="100%"
-              objectFit="cover"
-            />
+            <PostImage post={post} />
             <Box
               position="absolute"
               backgroundColor="primary"
